requirements.txt
landsatxplore
matplotlib
numpy~=1.19.2
opencv-python
Pillow
pygeoapi
rasterio
requests
scikit-image
scikit-learn
tensorflow~=2.6.0
tinydb
